+++ dose/about.py
@@ -1,9 +1,9 @@
 """Dose GUI for TDD: about dialog box."""
-import wx
 from . import __version__, __url__
 from .rest import single_line_block
 from .misc import snake2ucamel
 from .shared import README, CONTRIBUTORS, LICENSE
+from .compat import wx
 
 
 metadata = {
@@ -19,7 +19,7 @@ metadata = {
 
 def about_box():
     """A simple about box dialog using the distribution data files."""
-    about_info = wx.AboutDialogInfo()
+    about_info = wx.adv.AboutDialogInfo()
     for k, v in metadata.items():
         setattr(about_info, snake2ucamel(k), v)
-    wx.AboutBox(about_info)
+    wx.adv.AboutBox(about_info)